fix: change to use size_t for sign-compare warning

+++ src/hnswlib/hnswalg.h
@@ -739,7 +739,7 @@ namespace hnswlib {
             size_t dim = *((size_t *) dist_func_param_);
             std::vector<data_t> data;
             data_t* data_ptr = (data_t*) data_ptrv;
-            for (int i = 0; i < dim; i++) {
+            for (size_t i = 0; i < dim; i++) {
                 data.push_back(*data_ptr);
                 data_ptr += 1;
             }
@@ -1003,7 +1003,7 @@ namespace hnswlib {
                         unmarkDeletedInternal(existingInternalId);
                     }
                     updatePoint(data_point, existingInternalId, 1.0);
-                    
+
                     return existingInternalId;
                 }
